+++ lifelines/fitters/log_normal_fitter.py
@@ -2,27 +2,9 @@
 from __future__ import print_function
 from __future__ import division
 import autograd.numpy as np
-from scipy.stats import norm as _scipy_norm
 from autograd.scipy.stats import norm
-from autograd.extend import primitive, defvjp
-from autograd.numpy.numpy_vjps import unbroadcast_f
 from lifelines.fitters import KnownModelParametericUnivariateFitter
-
-
-logsf = primitive(_scipy_norm.logsf)
-
-defvjp(
-    logsf,
-    lambda ans, x, loc=0.0, scale=1.0: unbroadcast_f(
-        x, lambda g: -g * np.exp(norm.logpdf(x, loc, scale) - logsf(x, loc, scale))
-    ),
-    lambda ans, x, loc=0.0, scale=1.0: unbroadcast_f(
-        loc, lambda g: g * np.exp(norm.logpdf(x, loc, scale) - logsf(x, loc, scale))
-    ),
-    lambda ans, x, loc=0.0, scale=1.0: unbroadcast_f(
-        scale, lambda g: g * np.exp(norm.logpdf(x, loc, scale) - logsf(x, loc, scale)) * (x - loc) / scale
-    ),
-)
+from lifelines.utils.logsf import logsf
 
 
 class LogNormalFitter(KnownModelParametericUnivariateFitter):